backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pandas
strawberry-graphql
python-dotenv
langchain
langchain-groq
langchain-community
langgraph
requests
apscheduler
